First commit of the doxygen review.

+++ include/fastrtps/rtps/attributes/EndpointAttributes.h
@@ -19,7 +19,10 @@ namespace eprosima {
 namespace fastrtps{
 namespace rtps {
 
-	class EndpointAttributes
+/**
+ *
+ */
+class EndpointAttributes
 {
 public:
 	EndpointAttributes()
@@ -32,16 +35,43 @@ public:
 		endpointKind = WRITER;
 	};
 	virtual ~EndpointAttributes(){};
+	//!Endpoint kind
 	EndpointKind_t endpointKind;
+	//!Tipoc kind
 	TopicKind_t topicKind;
+	//!Reliability kind
 	ReliabilityKind_t reliabilityKind;
+	//!Durability kind
 	DurabilityKind_t durabilityKind;
+	//!Unicast locator list
 	LocatorList_t unicastLocatorList;
+	//!Multicast locator list
 	LocatorList_t multicastLocatorList;
+		
+	/**
+	 * Get the user defined ID
+	 * @return User defined ID
+	 */
 	inline int16_t getUserDefinedID() const {return m_userDefinedID;}
+	
+	/**
+	 * Get the entity defined ID
+	 * @return Entity ID
+	 */
 	inline int16_t getEntityID() const {return m_entityID;}
-	inline void setUserDefinedID(uint16_t id){m_userDefinedID = id;	};
-	inline void setEntityID(uint16_t id){m_entityID = id;	};
+	
+	/**
+	 * Set the user defined ID
+	 * @param User defined ID to be set
+	 */
+	inline void setUserDefinedID(uint8_t id){m_userDefinedID = id;	};
+	
+	/**
+	 * Set the entity ID
+	 * @param Entity ID to be set
+	 */
+	inline void setEntityID(uint8_t id){m_entityID = id;	};
+	
 private:
 	int16_t m_userDefinedID;
 	int16_t m_entityID;
